fix: use u64 for Money amount

+++ src/financial/money.rs
@@ -2,12 +2,11 @@
 ///
 /// ## Fields
 /// - `amount` - The amount of money.
-///    This is a floating point number to allow for fractional amounts.
 /// - `currency` - The currency of the money, represented as an ISO 4217 currency code.
 pub struct Money {
     amount: MoneyAmount,
     currency: Currency,
 }
 
-pub type MoneyAmount = f64;
+pub type MoneyAmount = u64;
 pub type Currency = String;
